requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
click==8.1.7
fastapi==0.108.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
iso8601==1.1.0
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
yarl==1.9.4
